+++ app/api/identities/get-identity.tsx
@@ -1,13 +1,7 @@
 import { cookies } from 'next/headers';
 import 'server-only';
 import { config } from "@/lib/logto-config";
-
-export interface IIdentity {
-  id: string;
-  name: string;
-  status: string;
-  flow: string;
-}
+import { IIdentity } from "app/api/interfaces/identity";
 
 export async function getIdentity() {
   const response = await fetch(`${config.baseUrl}/api/identities`, {
@@ -19,10 +13,8 @@ export async function getIdentity() {
 
   if (!response.ok) {
     if (response.status === 403) {
-      console.log(await response.json())
       return 'Access denied to method, requires read:identity scope.';
     }
-    console.log(response.status)
     throw new Error('Something went wrong!');
   }
