+++ app/api/interfaces/identity.ts
@@ -1,0 +1,8 @@
+export interface IIdentity {
+  handler: string;
+  displayName: string;
+  visibility: string;
+  bio: string;
+  active: boolean;
+  privacy: boolean;
+}
